backend/requirements.txt
supabase>=2.4.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
requests>=2.31.0
